perf: rewrite peripherals to use direct MMIO register access
Replaces dictionary-based read/write hook lookups with direct MmioController
register reads/writes across all peripherals (GPIO, SPI, TWI, USART, USI,
Timer, ADC, Clock, EEPROM, Watchdog). Removes three stale files that were
superseded by the Avr8Sharp versions (Adc, Clock, Eeprom under AVR8Sharp.Core).
